+++ packages/skeleton-svelte/src/lib/components/Popover/types.ts
@@ -6,6 +6,8 @@ export interface PopoverProps extends Omit<popover.Context, 'id' | 'open'> {
 	open?: boolean;
 	/** Enable display of the popover arrow. */
 	arrow?: boolean;
+	/** Disable the trigger element of the popover */
+	disabled?: boolean;
 
 	// Base ---
 	/** Set base classes for the root element. */
@@ -24,8 +26,8 @@ export interface PopoverProps extends Omit<popover.Context, 'id' | 'open'> {
 	// Positioner ---
 	/** Set base classes for the positioner. */
 	positionerBase?: string;
-	/** Set z-index classes for the positioner. */
-	positionerZIndex?: string;
+	/** Set z-index for the positioner. */
+	zIndex?: string;
 	/** Provide arbitrary classes for the positioner. */
 	positionerClasses?: string;